+++ frb_example/pure_dart/dart/lib/bridge_generated.web.dart
@@ -652,6 +652,10 @@ class FlutterRustBridgeExampleSingleBlockTestWasmModule implements WasmModule {
 
   external void wire_multiply_by_ten(NativePortType port_, List<dynamic> measure);
 
+  external void wire_call_old_module_system(NativePortType port_);
+
+  external void wire_call_new_module_system(NativePortType port_);
+
   external void wire_sum__method__SumWith(NativePortType port_, List<dynamic> that, int y, int z);
 
   external void wire_new__static_method__ConcatenateWith(NativePortType port_, String a);
@@ -827,6 +831,10 @@ class FlutterRustBridgeExampleSingleBlockTestWire
 
   void wire_multiply_by_ten(NativePortType port_, List<dynamic> measure) =>
       wasmModule.wire_multiply_by_ten(port_, measure);
+
+  void wire_call_old_module_system(NativePortType port_) => wasmModule.wire_call_old_module_system(port_);
+
+  void wire_call_new_module_system(NativePortType port_) => wasmModule.wire_call_new_module_system(port_);
 
   void wire_sum__method__SumWith(NativePortType port_, List<dynamic> that, int y, int z) =>
       wasmModule.wire_sum__method__SumWith(port_, that, y, z);