+++ frb_example/pure_dart/rust/src/new_module_system/sub_module.rs
@@ -1,0 +1,7 @@
+pub struct NewSimpleStruct {
+    pub field: i32,
+}
+
+pub fn use_new_module_system(value: i32) -> NewSimpleStruct {
+    NewSimpleStruct { field: value }
+}