+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.41.3
+
+* Fix "Skipping unresolvable module" by align latest #651 (thanks @alanlzhang)
+
+## 1.41.2
+
+* Add cli arg to skip dependencies check #640 (thanks @Roms1383)
+
 ## 1.41.1
 
 * Fix bug with conflicting names in enum name and variant #604 (thanks @Roms1383)